+++ src/server/routes/registerHooks.js
@@ -1,5 +1,7 @@
-import validateHook from '../utils/validateHook';
 import runCommand from '../process/runCommand';
+import areObjectValuesDefined from '../utils/areObjectValuesDefined';
+import pickVariablesFromJson from '../utils/pickVariablesFromJson';
+import validateHook from '../utils/validateHook';
 
 export default function registerHooks({ router, hooks }) {
     if (!router) throw new Error('No router provided');
@@ -19,9 +21,15 @@ function registerHook(params) {
     }
 }
 
-function addRoute({ router, method, path, command, cwd }) {
+function addRoute({ router, method, path, command, cwd, parseJson }) {
     router[method.toLowerCase()](path, (req, res) => {
-        const options = { cwd };
+        const env = pickVariablesFromJson(req.body, parseJson);
+
+        if (parseJson && !areObjectValuesDefined(env)) {
+            return res.status(400).send('JSON path does not match');
+        }
+
+        const options = { cwd, env };
         runCommand(command, options);
         res.json({ path, requestReceivedAt: Date.now() });
     });