+++ test/http/01-basic.js
@@ -44,17 +44,17 @@ describe('HTTP basic', () => {
             .end(handleHttpSuccess.bind(null, callback));
     });
 
-    it('GET /test3', (callback) => {
+    it('GET /test99', (callback) => {
         request(app)
-            .get('/test3')
+            .get('/test99')
             .set('Accept', 'application/json')
             .expect(404)
             .end(handleHttpError.bind(null, callback));
     });
 
-    it('POST /test3', (callback) => {
+    it('POST /test99', (callback) => {
         request(app)
-            .post('/test3')
+            .post('/test99')
             .set('Accept', 'application/json')
             .expect(404)
             .end(handleHttpError.bind(null, callback));